(refactor) Inline
